split large tables in several messages
